+++ source/renderer/app/themes/daedalus/white.js
@@ -373,12 +373,12 @@ export const WHITE_THEME_OUTPUT = {
       '#f9f9f94d',
     '--theme-legacy-notification-learn-more-button-background-color-active':
       '#f9f9f966',
-    '--theme-legacy-notification-move-button-text-color': '#ab2712',
+    '--theme-legacy-notification-move-button-text-color': '#ab1700',
     '--theme-legacy-notification-move-button-background-color': '#fafbfc',
     '--theme-legacy-notification-move-button-background-color-hover':
-      '#f9f9f9e6',
+      'rgb(250, 251, 252, 0.9)',
     '--theme-legacy-notification-move-button-background-color-active':
-      '#f9f9f9cc',
+      'rgb(250, 251, 252, 0.8)',
     '--theme-legacy-notification-title-color': '#fafbfc',
     '--theme-legacy-notification-description-color': '#fafbfc',
     '--theme-notification-message-checkmark-icon-color': '#fafbfc',
@@ -607,6 +607,25 @@ export const WHITE_THEME_OUTPUT = {
       'rgba(45, 45, 45, 0.3)',
     '--theme-sidebar-category-networkInfo-background-color': '#eb2256',
     '--theme-sidebar-category-networkInfo-text-color': '#121326',
+  },
+  splash: {
+    '--theme-splash-network-background-color': 'rgba(255, 255, 255, 0.96)',
+    '--theme-splash-network-background-color1': 'rgba(18, 19, 38, 1)',
+    '--theme-splash-network-background-color2': 'rgba(18, 19, 38, 0.91)',
+    '--theme-splash-network-background-color3': 'rgba(18, 19, 38, 0.31)',
+    '--theme-splash-network-background-color4': 'rgba(18, 19, 38, 0)',
+    '--theme-splash-network-background-color5': 'rgba(0, 0, 0, 0)',
+    '--theme-splash-network-background-color6': 'rgba(18, 19, 38, 0.31)',
+    '--theme-splash-network-background-color7': 'rgba(18, 19, 38, 0.91)',
+    '--theme-splash-network-background-color8': 'rgba(18, 19, 38, 1)',
+    '--theme-splash-network-logo-fill-color': '#29b595',
+    '--theme-splash-network-title-color': '#2d2d2d',
+    '--theme-splash-network-subTitle1-color': 'rgba(45, 45, 45, 0.5)',
+    '--theme-splash-network-subTitle2-color': 'rgba(45, 45, 45, 0.5)',
+    '--theme-splash-network-description-background-color':
+      'rgba(41, 181, 149, 0.1)',
+    '--theme-splash-network-description-color': 'rgba(45, 45, 45, 0.7)',
+    '--theme-splash-network-learn-more-color': '#2d2d2d',
   },
   stakePools: {
     '--theme-staking-stake-pools-title-color': '#2d2d2d',